+++ torch/_functorch/aot_autograd.py
@@ -5,7 +5,7 @@ import itertools
 import time
 from contextlib import nullcontext
 from functools import wraps
-from typing import Any, Literal, TYPE_CHECKING
+from typing import Any, cast, Literal, TYPE_CHECKING
 from typing_extensions import ParamSpec, TypeVar
 from unittest.mock import patch
 
@@ -161,7 +161,7 @@ if TYPE_CHECKING:
     from torch._inductor.compile_fx import CompilerConfigExtra
     from torch._inductor.output_code import OutputCode
     from torch._inductor.utils import InputType
-    from torch._ops import OpOverload
+    from torch._ops import OperatorBase, OpOverload
     from torch.fx.experimental.symbolic_shapes import ShapeEnv
 
 _P = ParamSpec("_P")
@@ -1528,6 +1528,32 @@ def aot_compile_joint_with_descriptors(
     return unflattened_compiled_fn
 
 
+@contextlib.contextmanager
+def _aot_export_decomposition_context(
+    decompositions: dict[OpOverload, Callable[..., Any]] | None,
+) -> Any:
+    if decompositions is None:
+        yield decompositions
+        return
+
+    from torch.export.decomp_utils import (
+        _override_composite_implicit_decomp,
+        _split_decomp_table_to_cia_and_python_decomp,
+        CustomDecompTable,
+    )
+
+    decomp_table = (
+        decompositions.materialize()
+        if isinstance(decompositions, CustomDecompTable)
+        else cast("dict[OperatorBase, Callable[..., Any]]", dict(decompositions))
+    )
+    cia_to_decomp, python_decomp_table = _split_decomp_table_to_cia_and_python_decomp(
+        decomp_table
+    )
+    with _override_composite_implicit_decomp(cia_to_decomp):
+        yield python_decomp_table
+
+
 def aot_export_module(
     mod: nn.Module,
     args: Iterable[Any],
@@ -1780,14 +1806,15 @@ def aot_export_joint_simple(
         # Run under no_grad, so our tracing machinery only traces an inference graph.
         ctx = torch.no_grad
 
-    with ctx():
-        fx_g, metadata, in_spec, out_spec = _aot_export_function(
-            func,
-            args,
-            decompositions=decompositions,
-            trace_joint=trace_joint,
-        )
-        in_spec, _kw_in_spec = in_spec.children()
+    with _aot_export_decomposition_context(decompositions) as decompositions_for_aot:
+        with ctx():
+            fx_g, metadata, in_spec, out_spec = _aot_export_function(
+                func,
+                args,
+                decompositions=decompositions_for_aot,
+                trace_joint=trace_joint,
+            )
+            in_spec, _kw_in_spec = in_spec.children()
     # At this point, we can just directly return the (joint or inference graph) that we traced.
     # First though: a bunch of assertions to make sure that our graph doesn't require
     # any calling convention changes compared to the original function.
